+++ plugins/plugin-shiki/package.json
@@ -35,9 +35,9 @@
     "clean": "rimraf --glob ./lib ./*.tsbuildinfo"
   },
   "dependencies": {
-    "@shikijs/transformers": "^1.5.2",
+    "@shikijs/transformers": "^1.6.0",
     "nanoid": "^5.0.7",
-    "shiki": "^1.5.2"
+    "shiki": "^1.6.0"
   },
   "peerDependencies": {
     "vuepress": "2.0.0-rc.11"
