+++ plugins/plugin-shiki/src/node/markdown/index.ts
@@ -1,3 +1,4 @@
 export * from './highlightLines.js'
 export * from './lineNumbers.js'
 export * from './preWrapper.js'
+export * from './inlineCode.js'
