+++ plugins/plugin-shiki/src/node/shikiPlugin.ts
@@ -3,6 +3,7 @@ import { isPlainObject } from 'vuepress/shared'
 import { highlight } from './highlight.js'
 import {
   highlightLinesPlugin,
+  inlineCodePlugin,
   lineNumberPlugin,
   preWrapperPlugin,
 } from './markdown/index.js'
@@ -19,6 +20,7 @@ export const shikiPlugin = (options: ShikiPluginOptions = {}): Plugin => ({
       ...options,
     })
 
+    md.use(inlineCodePlugin, options.vPre?.inline ?? true)
     md.use(highlightLinesPlugin)
     md.use(preWrapperPlugin, options)
     if (options.preWrapper ?? true) {
